refactor(admob): split tracking internals and reward verification scaffolding (#3470)
### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
Tracking APIs in AdMob are already shipped and should remain
compatibility-safe. This refactor improves internal structure by
separating tracking internals from the root package while keeping
reward-verification APIs as simple unimplemented placeholders.

### Description
This PR is **structure-only** in `feature/admob` (no intended behavior
change):

- Moved tracking internals into a dedicated internal area:
  - `.../admob/tracking/TrackIfConfigured.kt`
  - `.../admob/tracking/AdMobPrecisionMapping.kt`
  - `.../admob/tracking/TrackingAdListener.kt`
  - `.../admob/tracking/TrackingOnPaidEventListener.kt`
  - `.../admob/tracking/TrackingFullScreenContentCallback.kt`
- Kept public extension API files as thin entrypoints and rewired
imports:
  - tracking entrypoints remain in place and unchanged in signature
- Kept reward-verification extension APIs as direct
`NotImplementedError` stubs (no extra internal facade layer).
- Updated tests to reference tracking symbols from the new package.
- Removed old duplicated root-level tracking internal files.

**Before**
- Flat root-level package with tracking internals mixed with public
extension files.

**After**
- Root-level public API entrypoints remain thin.
- Tracking internals are grouped under `admob.tracking`.
- Reward-verification APIs remain unimplemented placeholders, unchanged
in behavior.

**Behavior parity statement**
- Tracking behavior and public tracking API are intended to remain
unchanged.
- Reward-verification public APIs remain unimplemented as before (same
`NotImplementedError` behavior).

### Test plan
- Attempted:
  - `./gradlew :feature:admob:test :feature:admob:lint`
- `./gradlew :feature:admob:testDefaultsBc8DebugUnitTest
:feature:admob:testDefaultsBc7DebugUnitTest`
- Blocked by unrelated pre-existing compilation failures in
`:purchases`:
-
`purchases/src/main/kotlin/com/revenuecat/purchases/interfaces/GetRewardVerificationStatusCallback.kt`
- error: internal RevenueCat API annotation usage violation during
`:purchases:compileDefaultsBc[7|8][Debug|Release]Kotlin`

No `feature/admob`-specific runtime/lint failures were surfaced before
hitting that blocker.
