+++ feature/admob/src/main/kotlin/com/revenuecat/purchases/admob/AdTrackerAdMobExtensions.kt
@@ -20,6 +20,9 @@ import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback
 import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAd
 import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAdLoadCallback
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.admob.tracking.TrackingFullScreenContentCallback
+import com.revenuecat.purchases.admob.tracking.setUpPaidEventTracking
+import com.revenuecat.purchases.admob.tracking.trackIfConfigured
 import com.revenuecat.purchases.ads.events.AdTracker
 import com.revenuecat.purchases.ads.events.types.AdFailedToLoadData
 import com.revenuecat.purchases.ads.events.types.AdFormat
