+++ feature/admob/src/main/kotlin/com/revenuecat/purchases/admob/AdViewExtensions.kt
@@ -10,6 +10,8 @@ import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.AdView
 import com.google.android.gms.ads.OnPaidEventListener
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.admob.tracking.TrackingAdListener
+import com.revenuecat.purchases.admob.tracking.TrackingOnPaidEventListener
 import com.revenuecat.purchases.ads.events.types.AdFormat
 import kotlin.jvm.JvmSynthetic
