+++ feature/admob/src/main/kotlin/com/revenuecat/purchases/admob/NativeAdExtensions.kt
@@ -11,6 +11,9 @@ import com.google.android.gms.ads.OnPaidEventListener
 import com.google.android.gms.ads.ResponseInfo
 import com.google.android.gms.ads.nativead.NativeAd
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.admob.tracking.TrackingAdListener
+import com.revenuecat.purchases.admob.tracking.TrackingOnPaidEventListener
+import com.revenuecat.purchases.admob.tracking.trackIfConfigured
 import com.revenuecat.purchases.ads.events.types.AdFormat
 import com.revenuecat.purchases.ads.events.types.AdLoadedData
 import com.revenuecat.purchases.ads.events.types.AdMediatorName
