+++ feature/admob/src/main/kotlin/com/revenuecat/purchases/admob/tracking/TrackingAdListener.kt
@@ -1,6 +1,4 @@
-@file:OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
-
-package com.revenuecat.purchases.admob
+package com.revenuecat.purchases.admob.tracking
 
 import com.google.android.gms.ads.AdListener
 import com.google.android.gms.ads.LoadAdError
@@ -26,6 +24,7 @@ import com.revenuecat.purchases.ads.events.types.AdOpenedData
  * Revenue is tracked separately via [com.google.android.gms.ads.OnPaidEventListener],
  * which must be wired on the ad view by the caller.
  */
+@OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
 internal class TrackingAdListener(
     internal val delegate: AdListener?,
     private val adFormat: AdFormat,
