+++ feature/admob/src/test/kotlin/com/revenuecat/purchases/admob/flow/BannerAdFlowTest.kt
@@ -8,6 +8,8 @@ import com.google.android.gms.ads.AdValue
 import com.google.android.gms.ads.AdView
 import com.google.android.gms.ads.OnPaidEventListener
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.admob.tracking.TrackingAdListener
+import com.revenuecat.purchases.admob.tracking.TrackingOnPaidEventListener
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
