+++ feature/admob/src/test/kotlin/com/revenuecat/purchases/admob/unit/AdMobPrecisionMappingTest.kt
@@ -4,6 +4,7 @@ package com.revenuecat.purchases.admob
 
 import com.google.android.gms.ads.AdValue
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
+import com.revenuecat.purchases.admob.tracking.toAdRevenuePrecision
 import com.revenuecat.purchases.ads.events.types.AdRevenuePrecision
 import org.junit.Assert.assertEquals
 import org.junit.Test
